Protofix: protocol hardening, robustness fixes, and test improvements
Tightens GENI frame/codec handling and improves overall component robustness:

**Protocol/codec hardening**
- Explicit `uint32_t`/`uint16_t` casts on BE decoders to avoid signed-shift UB
- Frame parser rejects truncated frames and clamps trailing-byte frames before CRC/payload extraction
- `frame_builder` rejects oversized SET payloads and GENI frames
- Route opspec-0x09 alarm/warning responses by echoed register address

**Lifetime/robustness**
- `Session::last_error_` is now `std::string` to avoid dangling pointer hazard
- Scheduler sequence counters in `Authentication` and `BLEConnectionManager` drop stale lambdas
- `Transport::loop` guards against missing write callback
- `DeviceInfoService` tracks per-string failures and reports aggregated success
- Auth extension packet naming/order aligned with `connection.md` Step C (wire order unchanged)

**DHW demand**
- `compute_deriv_` preserves both `prev` and `prev_ms` on NaN so next valid reading uses the correct elapsed time across any NaN gap
- `flow_latch_active_` derives sample count from actual update interval
- Charge-drop sets `onset_corroborating_signal_present`

**Tests & lint**
- Added BE uint16/uint32 helpers and signed-shift UB note to `protocol.h`
- Added negative-case CRC/parse tests; renamed test function to accurately reflect scope
- Extended cppcheck and clang-tidy targets to `dhw_demand` and `tests`

+++ components/alpha_hwr/codec.cpp
@@ -71,8 +71,10 @@ void encode_float_be(float value, uint8_t *bytes) {
 // Float decoding (big-endian)
 // Reference: alpha_hwr/protocol/codec.py::decode_float_be()
 float decode_float_be(const uint8_t *data, size_t offset) {
-  uint32_t temp = (data[offset] << 24) | (data[offset + 1] << 16) |
-                  (data[offset + 2] << 8) | data[offset + 3];
+  uint32_t temp = (static_cast<uint32_t>(data[offset]) << 24) |
+                  (static_cast<uint32_t>(data[offset + 1]) << 16) |
+                  (static_cast<uint32_t>(data[offset + 2]) << 8) |
+                  static_cast<uint32_t>(data[offset + 3]);
   float value;
   memcpy(&value, &temp, 4);
   return value;
@@ -88,7 +90,8 @@ void encode_uint16_be(uint16_t value, uint8_t *bytes) {
 // Uint16 decoding (big-endian)
 // Reference: alpha_hwr/protocol/codec.py::decode_uint16_be()
 uint16_t decode_uint16_be(const uint8_t *data, size_t offset) {
-  return (data[offset] << 8) | data[offset + 1];
+  return (static_cast<uint16_t>(data[offset]) << 8) |
+         static_cast<uint16_t>(data[offset + 1]);
 }
 
 // Uint32 encoding (big-endian)
@@ -103,8 +106,10 @@ void encode_uint32_be(uint32_t value, uint8_t *bytes) {
 // Uint32 decoding (big-endian)
 // Reference: alpha_hwr/protocol/codec.py::decode_uint32_be()
 uint32_t decode_uint32_be(const uint8_t *data, size_t offset) {
-  return (data[offset] << 24) | (data[offset + 1] << 16) |
-         (data[offset + 2] << 8) | data[offset + 3];
+  return (static_cast<uint32_t>(data[offset]) << 24) |
+         (static_cast<uint32_t>(data[offset + 1]) << 16) |
+         (static_cast<uint32_t>(data[offset + 2]) << 8) |
+         static_cast<uint32_t>(data[offset + 3]);
 }
 
 }  // namespace protocol
